sam_inpainting_app/requirements.txt
torch
opencv-python
numpy
matplotlib
Pillow
segment-anything
diffusers
transformers
accelerate
simple-lama-inpainting
scikit-image
streamlit
